+++ src/tr1/game/option/option_passport.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include <libtrx/game/inventory_ring/types.h>
-#include <libtrx/game/option/passport.h>